requirements.txt
# Core Framework
streamlit>=1.28.0

# API Client - OpenAI only
openai>=1.3.0

# Data Management - Updated for Python 3.13 compatibility
pandas>=2.2.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2